post func newuser and make parking subcollection
