--- Screens/SignUp.js
@@ -1,122 +1,0 @@
-import React, { useState } from "react";
-import { useNavigation } from "@react-navigation/native";
-// import { createUserWithEmailAndPassword } from "firebase/auth";
-
-import {
-    KeyboardAvoidingView,
-    StyleSheet,
-    Text,
-    TextInput,
-    TouchableOpacity,
-    TouchableWithoutFeedback,
-    View,
-    Keyboard,
-    Platform,
-} from "react-native";
-
-
-export default function SignUp() {
-    const [email, setEmail] = useState("");
-    const [username, setUsername] = useState("");
-    const [password, setPassword] = useState("");
-
-    const navigation = useNavigation();
-
-    const handleSignUp = () => {
-        // createUserWithEmailAndPassword(auth, email, password)
-        //     .then((userCredential) => {
-        //         const user = userCredential.user;
-        //         console.log("Registered with:", user.email);
-        navigation.navigate("HomeScreen");
-        // })
-        // .catch((error) => alert(error.message));
-    };
-
-    return (
-        <KeyboardAvoidingView
-            behavior={Platform.OS === "*" ? "padding" : "height"}
-            style={styles.container}
-        >
-            <Text style={styles.heading}>Sign up</Text>
-            <TouchableWithoutFeedback onPress={Keyboard.dismiss}>
-                <View style={styles.inputContainer}>
-                    <TextInput
-                        placeholder="Username"
-                        value={username}
-                        onChangeText={(text) => setUsername(text)}
-                        style={styles.Input}
-                    />
-
-                    <TextInput
-                        placeholder="Email"
-                        value={email}
-                        onChangeText={(text) => setEmail(text)}
-                        style={styles.Input}
-                    />
-                    <TextInput
-                        placeholder="Password"
-                        value={password}
-                        onChangeText={(text) => setPassword(text)}
-                        style={styles.Input}
-                        secureTextEntry
-                    />
-                </View>
-            </TouchableWithoutFeedback>
-            <View style={styles.buttonContainer}>
-                <TouchableOpacity
-                    onPress={handleSignUp}
-                    style={[styles.button, styles.buttonOutline]}
-                >
-                    <Text style={styles.buttonOutlineText}>Register</Text>
-                </TouchableOpacity>
-            </View>
-        </KeyboardAvoidingView>
-    );
-}
-
-const styles = StyleSheet.create({
-    container: {
-        fontSize: 53,
-        flex: 1,
-        justifyContent: "center",
-        alignItems: "center",
-        backgroundColor: "black",
-    },
-    inputContainer: {
-        width: "80%",
-    },
-    Input: {
-        backgroundColor: "white",
-        paddingHorizontal: 15,
-        paddingVertical: 10,
-        borderRadius: 10,
-        marginTop: 5,
-    },
-    buttonContainer: {
-        width: "60%",
-        justifyContent: "center",
-        alignContent: "center",
-        marginTop: 30,
-    },
-    button: {
-        alignItems: "center",
-        backgroundColor: "white",
-        width: "100%",
-        padding: 10,
-        borderRadius: 10,
-    },
-    buttonOutline: {
-        borderWidth: 2,
-        borderColor: "white",
-    },
-    buttonOutlineText: {
-        color: "black",
-        fontSize: 18,
-        fontWeight: "bold",
-    },
-    heading: {
-        color: "white",
-        fontSize: 20,
-        margin: 20,
-    }
-});
